% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCRi-package.R
\docType{package}
\name{BCRi-package}
\alias{BCRi}
\alias{BCRi-package}
\title{BCRi: Analysis of repertoire diversity across B cell phenotypes from single cell BCR-seq}
\description{
Analysis of repertoire diversity across B cell phenotypes from single cell BCR-seq
}
\author{
\strong{Maintainer}: Ashley Doane Stephen \email{ashley.doane@gmail.com} (\href{https://orcid.org/0000-0002-1489-1786}{ORCID})

}
\keyword{internal}
