% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-definitions.R
\docType{class}
\name{distance-class}
\alias{distance-class}
\title{distance-class}
\description{
Container for class \code{distance}.
}
\section{Fields}{

\describe{
\item{\code{distance}}{two-dimensional \code{matrix} of mode \code{numeric}
with rows as types, columns as types, and elements containing the pairwise
distance of types}

\item{\code{dat_id}}{object of class \code{character} describing the class of
distance / similarity being used, e.g. "naive", "taxonomic", and so on}

\item{\code{components}}{list containing the components necessary to calculate
similarity. This list is empty when \code{precompute_dist = TRUE} when
calculating distance. When a pairwise distance matrix is too large and
\code{precompute_dist = FALSE}, this list contains all the information
required to calculate pairwise distance between types}
}}

