% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiversityFunctions.R
\docType{package}
\name{DiversityFunctions}
\alias{DiversityFunctions}
\title{BCRi Diversity Functions}
\description{
This file contains the main diversity calculation functions for the BCRi package.
The package has been refactored into multiple files for better organization:
}
\details{
\itemize{
\item utils-diversity.R: Utility functions for validation and calculations
\item class-definitions.R: S4 class definitions
\item constructors.R: Constructor functions for objects
\item diversity-components.R: Low-level diversity components
\item subcommunity-diversity.R: Subcommunity-level diversity functions
\item metacommunity-diversity.R: Metacommunity-level diversity functions
\item generic-methods.R: Generic methods and S4 method definitions
\item repartition.R: Functions for repartitioning metacommunities
}
}
\author{
\strong{Maintainer}: Ashley Doane Stephen \email{ashley.doane@gmail.com} (\href{https://orcid.org/0000-0002-1489-1786}{ORCID})

}
