% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_diversity.R
\name{functional_diversity}
\alias{functional_diversity}
\title{Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone}
\usage{
functional_diversity(
  db,
  groupID = NULL,
  phenotype_var = "subset",
  phenotype_reference = NULL,
  cell_id = NULL,
  similarity = TRUE,
  group = "subject_id",
  cdr3 = FALSE,
  useAffinityWeights = TRUE,
  computeMax = FALSE,
  qs = 0:2,
  log_base = exp(1),
  distanceCutoff = FALSE,
  discreteVar = "clone_id",
  nboot = 100
)
}
\arguments{
\item{db}{db}

\item{phenotype_var}{phenotype variable}

\item{cell_id}{cell id}

\item{log_base}{base of log to use for shannon entropy calculations}

\item{cloneID}{ID of clone to analyze}

\item{clone}{column name of clone variable in db}
}
\value{
data.frame of diversity metrics and phenotype ##list with affinity matrix, db for clone, and data.frame of per-phenotype diversity metrics using Simposon diversity
}
\description{
Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone
}
