% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{getLevD}
\alias{getLevD}
\title{function to caclulate levenshtein distance matrix from a repertoire db}
\usage{
getLevD(
  db,
  region = c("sequence", "junction"),
  indVar = "ind",
  germline = "germline_alignment",
  sequence = "sequence_alignment",
  junction = "junction",
  cdr3 = FALSE
)
}
\value{
list of distance matrix and db with indicator variable for unique sequences
}
\description{
function to caclulate levenshtein distance matrix from a repertoire db
}
