% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{global_entropies}
\alias{global_entropies}
\title{function calculates the global entropy of the sequences in the db pver marginal phenotypes}
\usage{
global_entropies(
  db,
  region = "sequence",
  indVar = "ind",
  phenotype_var = "predictions_scanvi",
  germline = "germline_alignment",
  sequence = "sequence_alignment",
  junction = "junction",
  v_call = "v_call",
  j_call = "j_call",
  fields = NULL,
  locus = "locus",
  only_heavy = TRUE,
  split_light = FALSE,
  targeting_model = shazam::HH_S5F,
  len_limit = NULL,
  first = FALSE,
  cdr3 = FALSE,
  mod3 = FALSE,
  max_n = 0,
  threshold = 1,
  base_sim = 0.95,
  iter_max = 1000,
  nstart = 1000,
  nproc = 4,
  verbose = FALSE,
  log = NULL
)
}
\description{
function calculates the global entropy of the sequences in the db pver marginal phenotypes
}
