% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraclonal_diversity.R
\name{intraclonal_diversity}
\alias{intraclonal_diversity}
\title{function returns intraclonal diversity metrics using simpson (default) or shannon diversities
calculated using a BCR sequence similarity matrix}
\usage{
intraclonal_diversity(
  db,
  cloneID,
  use_clones = TRUE,
  method = c("simpson", "shannon"),
  phenotype_var = "subset",
  cell_id = NULL,
  clone = "clone_id"
)
}
\arguments{
\item{db}{db}

\item{cloneID}{clone to analyze}

\item{phenotype_var}{phenotype variable}

\item{cell_id}{cell id}

\item{clone}{column name of clone id in db}
}
\value{
list with affinity matrix, db for clone, and data.frame of per-phenotype
list with affinity matrix, db for clone, and data.frame of per-phenotype diversity metrics using using Simposon (default) or Shannon diversity calculated using a BCR sequence similarity matrix
}
\description{
function returns intraclonal diversity metrics using simpson (default) or shannon diversities
calculated using a BCR sequence similarity matrix
}
