% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraclonal_diversity.R
\name{intraclonal_shannon}
\alias{intraclonal_shannon}
\title{Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone}
\usage{
intraclonal_shannon(
  db,
  cloneID,
  phenotype_var = "subset",
  cell_id = NULL,
  clone = "clone_id",
  callClones = FALSE,
  normalize = "len",
  germline = "germline_alignment",
  sequence = "sequence_alignment",
  junction = "junction",
  v_call = "v_call",
  j_call = "j_call",
  fields = NULL,
  locus = "locus",
  only_heavy = TRUE,
  split_light = FALSE,
  targeting_model = NULL,
  len_limit = NULL,
  first = FALSE,
  cdr3 = FALSE,
  mod3 = FALSE,
  max_n = 0,
  threshold = NA,
  base_sim = 0.95,
  iter_max = 1000,
  nstart = 1000,
  nproc = 4,
  verbose = FALSE,
  log = NULL,
  summarize_clones = FALSE
)
}
\arguments{
\item{db}{db}

\item{cloneID}{specific clone}

\item{phenotype_var}{phenotype variable}

\item{cell_id}{cell id}

\item{clone}{column name for clone}
}
\value{
list with affinity matrix, db for clone, and data.frame of per-phenotype diversity metrics using Shannon diversity
}
\description{
Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone
}
