% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{intraclonal_simpson}
\alias{intraclonal_simpson}
\title{Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone}
\usage{
intraclonal_simpson(
  db,
  cloneID = NULL,
  phenotype_var = "subset",
  cell_id = NULL,
  clone = "clone_id",
  cdr3 = FALSE
)
}
\arguments{
\item{db}{db}

\item{cloneID}{ID of clone to analyze}

\item{phenotype_var}{phenotype variable}

\item{cell_id}{cell id}

\item{clone}{column name of clone variable in db}
}
\value{
data.frame of diversity metrics and phenotype ##list with affinity matrix, db for clone, and data.frame of per-phenotype diversity metrics using Simposon diversity
}
\description{
Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone
}
