% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraclonal_diversity.R
\name{pairwiseMutions}
\alias{pairwiseMutions}
\title{Helper Function for pairwise mutations, adapted from scoper}
\usage{
pairwiseMutions(
  germ_imgt,
  seq_imgt,
  junc_length,
  len_limit = NULL,
  cdr3 = FALSE,
  mutabs = NULL,
  norm_fact = TRUE
)
}
\description{
Helper Function for pairwise mutations, adapted from scoper
}
