% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_clone.R
\name{prepare_clone}
\alias{prepare_clone}
\title{Prepare Clone}
\usage{
prepare_clone(
  db,
  junction = "junction",
  v_call = "v_call",
  j_call = "j_call",
  first = FALSE,
  cdr3 = FALSE,
  fields = NULL,
  cell_id = NULL,
  locus = "locus",
  only_heavy = TRUE,
  mod3 = FALSE,
  max_n = 0
)
}
\arguments{
\item{db}{db}

\item{junction}{junciotn seq}

\item{v_call}{v_call}

\item{j_call}{j_call}

\item{first}{first allele or union}

\item{cdr3}{cdr3 only}

\item{fields}{additional fields eg subject}

\item{cell_id}{cell id}

\item{locus}{locus column}

\item{only_heavy}{default true}

\item{mod3}{mod3 seq only}

\item{max_n}{max}
}
\value{
list
}
\description{
Prepare Clone
}
