% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{prepare_global}
\alias{prepare_global}
\title{Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone
Builds affinity matrix as disjoint subgraphs using spectral clustering or vdj alleles.}
\usage{
prepare_global(
  db,
  cloneID = NULL,
  phenotype_var = "subset",
  cell_id = NULL,
  clone = "clone_id",
  cdr3 = FALSE,
  useAffinityWeights = TRUE,
  log_base = exp(1),
  distanceCutoff = FALSE
)
}
\arguments{
\item{db}{db}

\item{cloneID}{ID of clone to analyze}

\item{phenotype_var}{phenotype variable}

\item{cell_id}{cell id}

\item{clone}{column name of clone variable in db}

\item{log_base}{base of log to use for shannon entropy calculations}
}
\value{
data.frame of diversity metrics and phenotype ##list with affinity matrix, db for clone, and data.frame of per-phenotype diversity metrics using Simposon diversity
}
\description{
Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone
Builds affinity matrix as disjoint subgraphs using spectral clustering or vdj alleles.
}
