% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repartition.R
\name{repartition_boot}
\alias{repartition_boot}
\title{Bootstrap repartition metacommunity}
\usage{
repartition_boot(jdmat, aff_mat, nboot, qs)
}
\arguments{
\item{jdmat}{matrix of community data}

\item{aff_mat}{affinity matrix for genetic similarity}

\item{nboot}{number of bootstrap iterations}

\item{qs}{vector of q values for diversity calculation}
}
\value{
Returns bootstrap results for diversity analysis
}
\description{
Performs bootstrap resampling on metacommunity data for diversity analysis.
}
