% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{seqDistance}
\alias{seqDistance}
\title{Sequence set distance}
\usage{
seqDistance(seqs, AA = FALSE)
}
\arguments{
\item{seqs}{A character list of the sequences for which the distance is to be calculated.}

\item{AA}{Logical (FALSE by default). If to calculate the distance based on the amino acid sequences.}
}
\value{
A \code{list} containing a  \code{matrix} of the computed distances between the alleles pairs and a vector with length adjusted sequences.
}
\description{
Calculates the distance between pairs of BCRs based on their aligned sequences.
The function assume the sequences are at an even length.
If not the function will pad the sequences with to the longest sequence length with Ns.
}
