% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{set_eq_seqDistance}
\alias{set_eq_seqDistance}
\title{Pads the sequences to the longest sequence length with Ns}
\usage{
set_eq_seqDistance(seqs)
}
\arguments{
\item{seqs}{A character list of the sequences for which the distance is to be calculated.}
}
\value{
A \code{list} containing a  \code{matrix} of the computed distances between the alleles pairs and a vector with length adjusted sequences.
}
\description{
Pads the sequences to the longest sequence length with Ns
}
