% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{weighted_diversity_splits}
\alias{weighted_diversity_splits}
\title{Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone}
\usage{
weighted_diversity_splits(
  db,
  grpID = NULL,
  phenotype_var = "subset",
  cell_id = NULL,
  grp = "subject_id",
  cdr3 = FALSE,
  useAffinityWeights = TRUE,
  similarity = TRUE,
  log_base = exp(1),
  distanceCutoff = FALSE,
  discreteVar = "clone_id"
)
}
\arguments{
\item{db}{db}

\item{phenotype_var}{phenotype variable}

\item{cell_id}{cell id}

\item{log_base}{base of log to use for shannon entropy calculations}

\item{cloneID}{ID of clone to analyze}

\item{clone}{column name of clone variable in db}
}
\value{
data.frame of diversity metrics and phenotype ##list with affinity matrix, db for clone, and data.frame of per-phenotype diversity metrics using Simposon diversity
}
\description{
Takes a db, a cloneID, and the name of a phenotype variable and returns the affinity matrix, db, and data.frame of per-phenotype diversity metrics for that clone
}
